use consistent naming and refactor setup_helper.generate_parameter_module
