+++ generate_parameter_library_py/generate_parameter_library_py/generate_python_module.py
@@ -36,13 +36,13 @@ import os
 from generate_parameter_library_py.parse_yaml import GenerateCode
 
 
-def run(output_file, yaml_file, validate_header=None):
+def run(output_file, yaml_file, validate_file=None):
     gen_param_struct = GenerateCode("python")
     output_dir = os.path.dirname(output_file)
     if not os.path.isdir(output_dir):
         os.makedirs(output_dir)
 
-    gen_param_struct.parse(yaml_file, validate_header)
+    gen_param_struct.parse(yaml_file, validate_file)
 
     code = str(gen_param_struct)
     with open(output_file, "w") as f:
@@ -51,19 +51,19 @@ def run(output_file, yaml_file, validate_header=None):
 
 def parse_args():
     parser = argparse.ArgumentParser()
-    parser.add_argument("output_cpp_header_file")
+    parser.add_argument("output_python_module_file")
     parser.add_argument("input_yaml_file")
-    parser.add_argument("validate_header", nargs="?", default="")
+    parser.add_argument("validate_file", nargs="?", default="")
     return parser.parse_args()
 
 
 def main():
     args = parse_args()
-    output_file = args.output_cpp_header_file
+    output_file = args.output_python_module_file
     yaml_file = args.input_yaml_file
-    validate_header = args.validate_header
+    validate_file = args.validate_file
 
-    run(output_file, yaml_file, validate_header)
+    run(output_file, yaml_file, validate_file)
 
 
 if __name__ == "__main__":
